requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-vision==0.39.0